Uso de parser
We just has used the parser library just to evaluete the operations that are receiving from display. The parser allow us to compare the expression that is within display so it can operate just like any other math expression. You only have to be sure you're introducing valid math values